requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
cryptography==44.0.0
distlib==0.3.9
Django==5.0
django-allauth==65.3.0
django-auditlog==3.0.0
django-summernote==0.8.20.0
filelock==3.16.1
identify==2.6.3
idna==3.10
nodeenv==1.9.1
oauthlib==3.2.2
pillow==10.4.0
platformdirs==4.3.6
pre_commit==4.0.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.8.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.3
virtualenv==20.28.0
webencodings==0.5.1
